cdk-pipelines-project/requirements.txt
aws-cdk-lib==2.69.0
constructs>=10.0.0,<11.0.0
cdk-nag